Mv node tests into separate file
